Split parser core helpers
